ECA-2007: make sure we check for null object everywhere before we do entityManager.remove(Object).
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@11170 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/ejb/ca/publisher/PublisherQueueSessionBean.java
@@ -105,7 +105,13 @@ public class PublisherQueueSessionBean implements PublisherQueueSessionRemote, P
     	}
     	try {
     		org.ejbca.core.ejb.ca.publisher.PublisherQueueData pqd = org.ejbca.core.ejb.ca.publisher.PublisherQueueData.findByPk(entityManager, pk);
-    		entityManager.remove(pqd);
+    		if (pqd == null) {
+    			if (log.isDebugEnabled()) {
+    				log.debug("Trying to remove queue data that does not exist: "+pk);
+    			}
+    		} else {
+        		entityManager.remove(pqd);    			
+    		}
 		} catch (Exception e) {
 			log.info(e);
 		}
